+++ hugr-cli/src/lib.rs
@@ -1,13 +1,18 @@
 //! Standard command line tools, used by the hugr binary.
 
-pub use clap::Parser;
+use clap::Parser;
 use clap_stdin::FileOrStdin;
-use hugr_core::{extension::ExtensionRegistry, Hugr, HugrView};
+use clap_verbosity_flag::Level;
+use clap_verbosity_flag::{InfoLevel, Verbosity};
 use thiserror::Error;
+
+use hugr_core::{extension::ExtensionRegistry, Hugr, HugrView};
+
 /// Validate and visualise a HUGR file.
 #[derive(Parser, Debug)]
 #[clap(version = "1.0", long_about = None)]
 #[clap(about = "Validate a HUGR.")]
+#[group(id = "hugr")]
 pub struct CmdLineArgs {
     input: FileOrStdin,
     /// Visualise with mermaid.
@@ -16,6 +21,9 @@ pub struct CmdLineArgs {
     /// Skip validation.
     #[arg(short, long, help = "Skip validation.")]
     no_validate: bool,
+    /// Verbosity.
+    #[command(flatten)]
+    verbose: Verbosity<InfoLevel>,
     // TODO YAML extensions
 }
 
@@ -38,7 +46,7 @@ pub const VALID_PRINT: &str = "HUGR valid!";
 
 impl CmdLineArgs {
     /// Run the HUGR cli and validate against an extension registry.
-    pub fn run(&self, registry: &ExtensionRegistry) -> Result<(), CliError> {
+    pub fn run(&self, registry: &ExtensionRegistry) -> Result<Hugr, CliError> {
         let mut hugr: Hugr = serde_json::from_reader(self.input.into_reader()?)?;
         if self.mermaid {
             println!("{}", hugr.mermaid_string());
@@ -46,9 +54,15 @@ impl CmdLineArgs {
 
         if !self.no_validate {
             hugr.update_validate(registry)?;
-
-            println!("{}", VALID_PRINT);
+            if self.verbosity(Level::Info) {
+                eprintln!("{}", VALID_PRINT);
+            }
         }
-        Ok(())
+        Ok(hugr)
+    }
+
+    /// Test whether a `level` message should be output.
+    pub fn verbosity(&self, level: Level) -> bool {
+        self.verbose.log_level_filter() >= level
     }
 }