+++ hugr-core/src/hugr/views/descendants.rs
@@ -9,7 +9,7 @@ use crate::hugr::HugrError;
 use crate::ops::handle::NodeHandle;
 use crate::{Direction, Hugr, Node, Port};
 
-use super::{check_tag, sealed::HugrInternals, HierarchyView, HugrView, RootTagged};
+use super::{check_tag, HierarchyView, HugrInternals, HugrView, RootTagged};
 
 type RegionGraph<'g> = portgraph::view::Region<'g, &'g MultiPortGraph>;
 
@@ -175,7 +175,7 @@ where
     }
 }
 
-impl<'g, Root> super::sealed::HugrInternals for DescendantsGraph<'g, Root>
+impl<'g, Root> super::HugrInternals for DescendantsGraph<'g, Root>
 where
     Root: NodeHandle,
 {