+++ hugr/src/lib.rs
@@ -134,9 +134,12 @@
 
 // These modules are re-exported as-is. If more control is needed, define a new module in this crate with the desired exports.
 // The doc inline directive is necessary for renamed modules to appear as if they were defined in this crate.
-pub use hugr_core::{builder, core, extension, hugr, ops, std_extensions, types, utils};
+pub use hugr_core::{builder, core, extension, ops, std_extensions, types, utils};
 #[doc(inline)]
 pub use hugr_passes as algorithms;
+
+// Modules with hand-picked re-exports.
+pub mod hugr;
 
 // Top-level re-exports for convenience.
 pub use hugr_core::core::{