+++ hugr-core/CHANGELOG.md
@@ -1,1 +1,17 @@
 # Changelog
+
+## 0.1.0 (2024-05-29)
+
+### Bug Fixes
+
+- Set initial version of hugr-core to 0.1.0 ([#1129](https://github.com/CQCL/hugr/pull/1129))
+
+### Features
+
+- [**breaking**] Move cli in to hugr-cli sub-crate ([#1107](https://github.com/CQCL/hugr/pull/1107))
+- Make internals of int ops and the "int" CustomType more public. ([#1114](https://github.com/CQCL/hugr/pull/1114))
+- Unseal and make public the traits `HugrInternals` and `HugrMutInternals` ([#1122](https://github.com/CQCL/hugr/pull/1122))
+
+### Refactor
+
+- Add a `hugr-core` crate ([#1108](https://github.com/CQCL/hugr/pull/1108))